requirements.txt
faiss-cpu
langchain
langchain-core
langchain-community
sentence-transformers
fastapi
uvicorn
requests
python-dotenv
openai
# duckduckgo-search
ddgs
chromadb